requirements.txt
requests
tqdm
python-dateutil